First trial translator.py
So far I have finish it and ran a couple of my own test cases, however using a lot of nested and a lot of if so if I have time try to shrink this as best as you can and make it more efficient. Also test more instances like number spaces and letter together.